requirements.txt
pyyaml
pyopml
beautifulsoup4
lxml